Move CovidTests Screens & Components (#501)
* Move Covid Tests into different folder

* Add changes to CovidApp
